Moved a function to a separate file.
